Further refactoring
