+++ pymt/Stop.py
@@ -1,4 +1,4 @@
-import parser
+from pymt import oasth_parser
 
 
 class Stop:
@@ -25,7 +25,7 @@ class Stop:
         # params['dir']
 
         if payload is not None:
-            parser.parse_stop(self, payload, uid)
+            oasth_parser.parse_stop(self, payload, uid)
 
         else:
             self.params = params
@@ -39,4 +39,4 @@ class Stop:
         self.buses.append(bus)
 
     def update(self, payload):
-        self.buses = parser.parse_stop_buses(payload)
+        self.buses = oasth_parser.parse_stop_buses(payload)
